+++ Source/CesiumEditor/Private/CesiumPanel.cpp
@@ -158,7 +158,7 @@ TSharedRef<SWidget> CesiumPanel::BasicQuickAddPanel() {
       "",
       -1});
   quickAddPanel->AddItem(QuickAddItem{
-      QuickAddItemType::CARTO_POLYGON,
+      QuickAddItemType::CARTOGRAPHIC_POLYGON,
       "Cesium Cartographic Polygon",
       "An actor that can be used to draw out regions for use with clipping or other material effects.",
       "",